Fixing TRUTH0 for running on HITS
With TRUTH0, we explicitly set up the converter (independent of the
standard MC truth tools setup).  We missed adding into TRUTH0 a handler
for the case of running on HITS, where the converter might be made
automatically and would not need to be set up.  This fixes that issue
and lets the derivation making run to completion.  It has no effect on
standard jobs.


Former-commit-id: 5483362659a4bf5adbac40f24cf35ca00db4041c