Use branch names rather then leaf names in TTree->HDF5 converter
The copyRootTree code behind the ttree2hdf converter uses the leaf name rather than the branch name when loading the contents from ROOT. This is necessary to get the type (which seems to be property of the leaf rather than the branch) but if it used as the key to get the branch contents, it is vulnerable to the case where the author of the TTree uses a different string for the branch name and the leaf name, e.g. if they do something like:

```
tree->Branch("muon_pT",&muon_pT,"b_muon_pt/F");
```

rather than the more usual convention:

```
tree->Branch("muon_pT",&muon_pT,"muon_pt/F");
```

In the first case above, the code as it currently stands will throw warnings about missing branches, and will then fill the HDF5 file with garbage (uninitialised values), presumably because all it can see is the type and number of entries, but not the numerical payload.

This MR makes a small modification, such that the keyword passed to the emplace_back method is now the name of the branch to which the leaf is associated, rather than the name of the leaf. The leaf name is still needed to get the type, so GetListOfLeaves is still used (in contrast to what we thought when we discussed this on GitHub).

Let me know what you think.

James


Former-commit-id: c2389e09d20811ecaefe3b4d438132390bf0e954