Change the name of a static variable
Apparently the naming of a static class variable did not conform to the ugly
atlas code convention regarding static variables.

The RoiDescriptor is used in pretty much every trigger class so modyfying this
will require essentially the entire trigger code base, (and quite some offline code)
to be compiled so needs to be submitted in the release before it can be properly
checked.

This addresses ATR-17619.


Former-commit-id: ac437b6ddf7c424a9a74969bcbb112b3a864254e