Separate the fullscan instance of the FTK moitoring
Separate the fullscan chains to their own instance of the monitoring.
This allows the track quality selection to be made independently for
these analyses.

In addition, I tidied the indentation - there were mixed tabs and spaces
and python was complaining, so I emoved them all. Unfortunately, this
makes it look as if the entire file has been updated because git diff
doesn't ignore white space, although there are only really about 15
lines which have actually changed.


Former-commit-id: b7ba787a481b33a18c92aa47e2ac2c428d684b4d