Fix the chain parser to allow FTK chains and fix rmain to allow general tau analyses
The descriptions say everything, not much mpre needs to be said - the tai analyses
which are not labeled as 1Prong or 3Prong were exclucded, not any analysis with Tau
can also be used.

The chainparser would fail for full scan FTK based analyses, because of the "chain name"
not starting with HLT etc. This fixes that.


Former-commit-id: e01f645b9294cadc3e54d9ef27f446c87d7eaac3