Fix hard-coded values in sTGC Readout Geometry (ATLASSIM-3675)
Some values used to build the sTGC readout geometry were hard-coded or
depended on hard coded values. This commit modifies the geometry xml and
adds the extra defining parameters as well as the AGDD tools required to
access the values and remove the bad coding practices.
Two new xml files were added with the same additions. The first,
stations.v1.78 is used as the soft-link. At the time of this commit, we
are in the transition between two geometry formats, but are still using
version1. Version2 will be used in the short future and as such the
changes were future-proofed.

Update NSW geometry soft link to versions.1.78

Add MuonAGDD tools for new sTGC parameters

Remove hard-coded values in sTGC Readout


Former-commit-id: bfa0b5b2ec28122e47ba0ad1f16ffbcfa2bc1b94