Fix rare crash for tau trigger element links
This MR implements a fix for element links in tau trigger.
In rare occasions, the newly implemented EL patch crashes (ATR-18077).
This happens when a tau reconstruction tool fails, the tau is then removed (and not recorded), but the code was still trying to fix the element links for that removed tau.
A simple protection has been added.
It is urgently needed in 21.1 and 21.1-dev.

Cheers,
Bertrand


Former-commit-id: bce229dda8872401cb234e0f4311848e97aefbf9