Allow the post processing to read in several GRLs for each job
PReviously only one GRL could be read in, meaning that the GRL had to be
specified differently for running over data from any given year.

This chage allows us to read in GRLs from several years in a single job
so that the same list will work for any year without needing to be configured
differently.


Former-commit-id: d5a656eb20aad5706760b391a25d16c2f08fcd5c