JetEventTPCnv: Avoid calling method through null pointer.
Backport from master:

There were a couple instances where a method was being called from
a null pointer.  The method didn't access any class member data, which is
why it didn't crash in the past.  However, this is undefined, and now
causes a crash when the undefined behavior sanitizer is used.


Former-commit-id: a386fa6453d088b4e7eb3b01f341ddfa6a4a0661