Fix AFP bytestream converter to work with HLT. ATR-17821
Bytestream converter was stopping when header information was
missing. This information is present only in offline, that is why no
AFP hits were read out in online. This is fixed by using the header
information only what it exists.

Additionally shift in x direction in AFP hypothesis algorithm was set
to 0, because after updates tracks are reconstructed in station
coordinate system i.e. around 0 and not -97.


Former-commit-id: fa9f437ada4fb4ee439665dd4b8bef38e4d70e2c