AthenaPython: Make PyAthenaGILStateEnsure.h public.
In an MT job, we need to always make sure we're holding the python GIL
before calling any python code.  There's a utility in this package
that does exactly that, but it was private to this package.
Make it public, so it can be used from other packages as well.


Former-commit-id: f46a2ee563f65f336e9fa7b474525b988c9cc942