cutting backward compatibility to very old layouts (ATLASSIM-3694)
In the past there were lots of different layouts for the muon
system. Already before data taking started in run-1 there was
convergence to MuonSpectrometer-R.XX.XX layouts. Furthermore,
the muon envelope is called since a very long time MUONQ02. This
MR cuts the backward compatibility to  other envelope names used
in the past and is part of the efforts to rename it to just MUON
eventually.


Former-commit-id: 2eaa08622097647c1f0fe2c00271805fac13c0b1