cut backward compatibility to Q02 in MuonSwitches (ATLASRECTS-2134)
In the past there were lots of different layouts for the muon
system. Already before data taking started in run-1 there was
convergence to MuonSpectrometer-R.XX.XX layouts. Each layout
has a version string in the attached table MuonSwitches. This
string became R.XX.XX while XX.XX corresponds to the version
number. Earlier strings like "Q02_initial" or just "Q02" are
not anymore content of the MuonSwitches table of any of the
supported layouts in release 21. Therefore the code can be
cleaned up and all related if-statements can be removed.
The parts corresponding to "Q..." are done in this MR. Still
left are "P..." and "C...".

A acmd.py diff-root comparison was performed on simhits of
2000 single muon events. The hits files before and after the
modification are identical.


Former-commit-id: 04e3e5912d64e25cf76ed72fcc2f6e089d17bc5b