cut backward compatibility to P03 layouts in CSC digitization
In the past there were lots of different layouts for the muon
system. Already before data taking started in run-1 there was
convergence to MuonSpectrometer-R.XX.XX layouts. Each layout
has a version string in the attached table MuonSwitches. This
string became R.XX.XX while XX.XX corresponds to the version
number. Earlier strings like "Q02_initial" or just "Q02" are
not anymore content of the MuonSwitches table of any of the
supported layouts in release 21. Therefore the code can be
cleaned up and all related if-statements can be removed.
This MR removes the compatibility to "P03" layouts in the
CSC digitization.

A acmd.py diff-root comparison was performed on digits of 25
ttbar events. The digit files before and after the modification
are identical (except for the timing leave of course).


Former-commit-id: 4ef41b7cb651289b2d52180a2eaed4d273f6f13f