cutting MGM compatibility to P-layouts for muons (ATLASRECTS-2134)
In the past there were lots of different layouts for the muon
system. Already before data taking started in run-1 there was
convergence to MuonSpectrometer-R.XX.XX layouts. Each layout
has a version string in the attached table MuonSwitches. This
string became R.XX.XX while XX.XX corresponds to the version
number. Earlier strings like "P03" or "CTB" are not anymore
content of the MuonSwitches table of any of the supported layouts
in release 21. Therefore the code can be cleaned up and all
related if-statements can be removed. This MR cleans from
"P.." and "C.." (also in the commented methods which themselves
should be kept though!).

A acmd.py diff-root comparison was performed on simhits of 2000
single muon events. The hits files before and after the
modification are identical (except for the time of course).

A acmd.py diff-root comparison was performed on digit output
of 25 ttbar events. The digit files before and after the
modification are identical (except for the time of course).


Former-commit-id: 02f0e1532328d2e9a6076377640d2680be8ac763