MuonGeoModel: Fix undefined behavior sanitizer warnings.
Backported from master:

We were getting warnings about casting to a SupComponent, where the
object was not actually a SupComponent.  But in fact, objects of
SupComponent are never created.  The only difference between SupComponent
and its base class Standard Component is the addition of a few methods
that use public class data.  So we can just as well not have SupComponent
derive from StandardComponent and change those methods to statics,
taking a StandardComponent as input.

Also removed Sup.h/Sup.cxx, which is not actually used anywhere.

We were getting warnings about casting to a SpacerComponent, where the
object was not actually a SpaComponent.  But in fact, objects of
SpacerComponent are never created.  SpaComponent and SpaComponent
are identical except for the name, so resolve this error by getting
rid of SpacerComponent (which never actually had any instances created)
and using SpaComponent instead.

Former-commit-id: cf5c3092e244fc8c4f50a56c917dac402c75b176