MuonReadoutGeometry: Fix undefined behavior sanitizer warning.
Invalid cross-cast: the code was casting from GeoVFullPhysVol to
GeoPhysVol, but one isn't a base of the other.  Looks like the cast
was meant to be to GeoFullPhysVol.

Invalid cross-cast: the code was casting from GeoVFullPhysVol to
GeoPhysVol, but one isn't a base of the other.  Looks like the cast
was meant to be to GeoFullPhysVol.

Former-commit-id: 4b5ccd23a38d5caa89cba4fdbff2b84897743a95