L1Decoder: Fix reference to uninitialized data.
If there aren't as many input fields in roi as parseInputRoI expects,
it'll return uninitialized data (which can lead to a practically infinite
loop when we try to normalize the phi).  Add an elementary check to
fix such a loop observed from creatingEVTest in ViewAlgsTest.
However, this should really be fixed to do proper input validation.


Former-commit-id: 4bb6745977f4680617f0b2f5811f4196dadb1204