Added a unit test for CP::detail::ShallowCopy, which would only build/run in standalone mode.
Since ElementLinks behave differently in the standalone and Athena builds, testing the shallow copying code can only be done in a build specific manner. And using the standalone event store for testing
is just a lot easier than using StoreGate. :-P

The test just makes sure that copying a view container is possible, and that the result of the copy is what we would expect.


Former-commit-id: 79dab14bebb9accf4823aeff204f0e13f9e07c96