Renamed the standalone IProxyDict typedef to AthContainers::IProxyDict.
This is to avoid ROOT getting upset about different types being declared as
IProxyDict during a job. As a more faithful emulation of this "offline" type
is in xAODRootAccess.

Unfortunately I couldn't just use the xAODRootAccess type in AthContainers,
as that does need the underlying xAOD::TVirtualEvent type. (As that's the type
expected by the standalone implementation of the ElementLink class.) So this
was the best that I could come up with...


Former-commit-id: 79dfba06a826fd65b35a899ef28ddc6e091755ca