Added an extra check in xAOD::TAuxStore::isContainer(...) to fix ATLASG-1392.
It seems that with the latest version of clang on macOS we can't depend purely
on the comparison of std::type_info instances in memory. It seems that multiple
instances can be assigned to the same type in certain situations.

So I added a final check, before the code would fail, that simply compares the
string names of the types to decide whether we are looking at a container or
a single object. This seems to fix the problem described in ATLASG-1392.


Former-commit-id: bc655158d3c68849464e6d5a9c80eb7cee5ea571