commenting unused variables like suggested on !11363
The shifter reviewing the sweep to master advised to resolve the
compiler warnings caused by the variables which are not longer
used after the last commit. They were accordingly commented on
the cherry-pick branch to master. I'm aware that commenting is
rather disfavored, however, for documentation purposes i would
appreciate if the lines could be kept.


Former-commit-id: 13dfe26bbf2c1a89f7161c0a67039638fa77b4f3