FTE fix caused by curvature parallel to surface (ATLASRECTS-4474)
Like described on the linked Jira ticket there are enormous
errors during track extrapolation in cases when the curvature
is parallel to the surface. @kluit provided a fix which is
content of this MR.

The frozen tier 0 test was passed (log file is available at
/afs/cern.ch/user/j/jomeyer/workarea/public/ATLASRECT_4474/RunTier0Tests.log )
even though obviously the fix violates frozen Tier0. The example
event on the ticket does now pass reco w/o FTE. Please @tvazquez
and @jbeacham have an eye on your ART tests when this MR goes
in. It would interesting to understand if you see any FT0
violating changes on your high stats tests.


Former-commit-id: a3d826f8c4e64cd82c3f930c84a20781e0dc4ef0