Adding catches for Zgamma/Wgamma for ATLASG-1433
Prompt Zgamma or Wgamma production, as well as ISR (basically the same
process but would be named differently by a generator) were not being
classified appropriately.  Added a fall-back to catch these cases at the
end of the classification routines, so that they won't interfere with
any of the other classifications, but can catch some of the edge cases
we weren't already dealing with.


Former-commit-id: b8b6e3977a32c18ac10487502bb084fba5a13097