Address the comments by L2 shifter
According to the comments from L2 shifter, the codes are adjusted to
follow the ATLAS guidelines. Some unnecessary commented lines are also
deleted.

An extra job option for the mass hypotheses of the particles in the
intermediate/final states is provided so that the algorithm would be a
bit more general.


Former-commit-id: 2e295cde7a034be019c2750debbde7b442890f0a