Protected TrigDecisionTool against not cleaning up after itself in standalone mode.
In order to avoid problems in standalone jobs that don't explicitly call finalize()
on the TDT, but which do delete the TDT object correctly, I copied part of the
cleanup code from the finalize() function into the Trig::TrigDecisionTool destructor
as well.

This should avoid the problems seen in some analyses when running on multiple datasets
in the same job with EventLoop.


Former-commit-id: 8c89cbe4b7c41521f44e4519674d4bdd8d062b1d