AthenaServices: Add TransientItems property to AthenaOutputStream.
We'd like to be able to have AthenaOutputStream properly declare
input dependencies.

This is a bit complicated because many configurations initialize the ItemList
with the full set of everything that could possibly be written.
If we simply make input dependencies for all of them, then we get
a massive set of warnings from SGInputLoader.  There's no good way
of suppressing these, since SGInputLoader doesn't have any information
on which algorithm was responsible for which dependency.

Instead, we add a new ItemList-like property, TransientItems, which
is meant to be filled with data items we know are present in the
transient store.  AthenaOutputStream then declares input dependencies
for everything in the item list that is also in TransientItems.

We'll arrange for TransientItems to be filled in in OutputStreamAthenaPool.py
from the Transient list in objKeyStore.  So to take advantage of this,
configuration code will need to record outputs in objKeyStore.


Former-commit-id: 28230bfbb36b1628ae78f198289c8d278cf1a3ad