Downgrade ERROR to WARNING in PathResolver for files in dev calib area
Currently, PathResolver issues a recoverable ERROR in case it find '/dev' in the path.
This is causing trouble in 21.1-dev where calibration files stored in the 'dev' area are used (ATR-18064).
This MR downgrades the ERROR to a WARNING, to avoid fake errors in git CI, and other ART tests.
Of course, this MR is meant only for 21.1-dev.


Former-commit-id: be1f1d71fc4ced0a1d5f5cbe3b1389f3feb702ba