Introduced a dual-use AlgSequence and AnaAlgSequence python class.
The AnaAlgorithm.AlgSequence class is just meant to emulate Athena's
AthSequencer class for EventLoop. (In Athena mode the user is given
the AthSequencer class when importing this code.)

Most of the development will go into the AnaAlgorithm.AnaAlgSequence
class, which is a common layer above the Athena and EventLoop implementations
of the algorithm sequence. Providing helper methods for correctly
configuring an analysis algorithm sequence.


Former-commit-id: fd409a8d2e85c8569d43c8b723d0dab70753121b