Migrated Muon Digit To RDO code to use Read/WriteHandles
The `MdtDigitToMdtRDO`, `RpcDigitToRpcRDO` and `TgcDigitToTgcRDO`
`AthAlgorithms` and the `CscDigitToCscRDOTool` have been migrated
to use `SG::Read/WriteHandleKeys` as member variables and to use
`SG::Read/WriteHandles` internally. This should mean that their
data dependencies are properly declared. Part of the changes for
ATLASSIM-2899.

Also remove usage of the obsolete `ActiveStoreSvc`.


Former-commit-id: 4cde9179d570f6dd9a04d93c696401d44add09af