Add job options to manually specify data-dependencies of Algorithms
In order for the Scheduler to correctly order Algorithms in MT running,
the data-dependencies of Algorithms need to be correctly specified.
This can be done by ensuring that they have `SG::Read/WriteHandleKey`
member variables for the objects that they intend to access in `StoreGate`.
For those `Algorithms` which haven't yet been migrated to use `SG::Read/WriteHandleKeys`
the dependencies can be specified using the `ExtraInputs` and `ExtraOutput`
properties in a postInclude. This is part of the changes for ATLASSIM-2899.


Former-commit-id: 06d4985f817928a7aa704fa1c5f1a0911c39b54b