Fix plot labelling and Shifter references in the comparitor
The new changes for the FTK chains and the plot panels screwed up the default
labeling in the keys and screwed up the automatic selection of the reference
in the Shifter chains.

These changes rework the way the keys are selected and the Shifter reference
determined.

Should now work for the Shifter, FTK Fullscan, and standard Expert type analyses


Former-commit-id: c6346998c129338ee52d7b9ddaed605e4bf574f6