Fix the atlas label position
because the TPad GetWw() and GetWh() methods return the size of the *TCanvas* rather
than the actual current gPad the calculation of the offset doesn't work for multiplot
canvases.

This allows the number of rows and columns of plots to be passed in so thatthe offset
can be calculated correctly.


Former-commit-id: 5ae169569ce5566bf4814ee8d3f7af4f7d15fd56