low beta examples in MuonCombinedRecExample and perf algs
There is a dedicated reco software for low beta particles which
traverse the muon spectrometer. To better test and more frequently
check this software a signal MC sample was copied to the CERN site
and the file list in MuonCombinedRec_myTopOptions.py got extended.
Furthermore the flag doMuGirlLowBeta is added in the python file
for more convenient enabling (default is False). To monitor the
reconstruction efficiency of the stau algorithm according instances
of the common muon performance algorithms are now created and
enabled if doMuGirlLowBeta is set true.


Former-commit-id: 2b89a6ce53450e771bb1d10625302da76ec5e81a