new dictionary for muon phase-2 layouts (w/o NSW)
Soon there will be a muon phase-2 layout on the database. This
layout will incorporate BIS78, BIS sMDT and full BI RPC. An
according dictionary for the IdHelpers is needed. At first,
the one coming with this MR, is just a carbon copy of R.04 for
BIS78. However, to not let the layout crash immediately and
on the other hand not interfere with other developments, the
simple copy is put in place.


Former-commit-id: ef7d19141fcee724b94b655bf47c681cf7c77bff