Generalised the addPrivateTool(...) function to work with sub-tools of tools.
It turned out that we need to use a slightly different logic when specifying
a "tool name" with periods in it.

In the end this generalisation was not needed, since the sub-tool of a tool that
was set up in TauAnalysisAlgorithms, turned out to be a public tool. Still,
I wanted to add this code, as it should provide some more robustness for our
future code.


Former-commit-id: 2a79227c3bd6165c5de590af9a89ca3332e5448a