Auto-configuring CASE when needed
We have a flag for R-hadron simulation, CASE, which describes whether
we're working in a model with stop, sbottom, or gluino R-hadrons.  Of
course, once you know the spectrum you should know which case you're in.
Added configuration code so that if the CASE isn't specified in the EVNT
metadata we look through the mass spectrum and try to figure out what
case we should be in, and set it automatically.  Will eventually allow
easier user configuration.


Former-commit-id: 758dd597846f150fc5ca6222cfdd187fd5c5e994