remnant adjustment for Run-3 geometries missed in !10658
The current layout for the NSW is marked as run-3 layout in the
AtlasCommon table. The information which run period a layout
corresponds to should be retrieved from this table. It should
by any means be avoided to check the layout name for such
information. Unfortunately there were still places in the
code which did that. This MR replaces the according lines
by the appropriate method. The previous if-statement also
took into account layouts with "IBL3D25" in the name. Such
layouts are not at all supported anymore in release 21, and
therefore the case was removed.


Former-commit-id: 577c72b0f3da8d940432028e0c9f044b9cc84e9d