SUSYTools: Add DCube post-processing for ART tests
Added dcube comparisons for our ART tests.  Small hack/workaround is needed
to trick the standalone dcube installation into thinking that the
release is Athena (since it doesn't yet know about the 21.2 projects).
This will be removed soon (the needed commits are already made, just need
to wait for those changes to be passed on to the standalone release).


Former-commit-id: 7fab0c540f829739cf6cb157067916893f3f9dde