Fix for bug in MuonTrackCleaner
When comparing the results of removing chambers, the MuonTrackCleaner makes a list of removed hits so that they can be added back when the removal of the next chamber is tested.
The problem is that the code was not checking to see if the hits had already been designated as outliers and so should not be added back.
The code uses its own map of the hits for this purpose, so this does not affect whether or not the hits are used in the fit.
Instead, it means that the code thinks that there are more hits available for fitting than there really are: in particular, that there are at least 6 (if there are fewer, it does not try to do a fit).
For sufficiently poor-quality tracks the code can end up trying to fit a track with one hit in the MS, and hence 0 dof: this leads to an FPE.
Both ATLASRECTS-4483 and ATLASRECTS-4487 report FPEs that result from this bug.
This fix resolves the problem, and removes the just-mentioned FPEs, by checking to make sure that hits were not originally outliers before marking them as being removed.
(Additionally, the debugging output is improved a bit.)
Some FT0 violation is presumably possible but it should be minimal and none is observed from the RunTier0Tests.


Former-commit-id: 4f5666015cc554667da5c30244b26b35caf7eedd