Removed redundent P03/H8 code (ATLASRECTS-4433)
The list of currently supported layouts does not include
any layout of type P03 or H8. The oldest layout is of type R06
(e.g. in ATLAS-R1-2010-02-00-00). However, the RPC cabling still
supports P03/H8. This backward compatibility should be cut in
21.3 and further. The clean-up can help during developments
of the package (e.g. in view of BIS78) and make it more lucid.


Former-commit-id: d56a98fa01d2b9eba7d97c17f0e851c52def8831