Tweaked TauAnalysisTools::CommonEfficiencyTool to avoid histogram mis-casts.
Changed the code to handle histograms through the simplest pointers possible.
To avoid mis-casting histograms, which was indeed happening according to GCC.
At the same time changed ReadInputs(...) to receive a simple reference to a
TFile object instead of a reference to a unique pointer. (Which is not a good
interface for a function...)


Former-commit-id: 765b9b85d45a35a229e254baab3cede82a29e559