SUSYTools: Add functionality for jet cleaning variablesin DAODs
By default SUSYTools will now query to see if the input file has
the jet cleaning flag already filled.  If so, it will not use the
jet cleaning tool.  There is a fallback available, just in case
the user is running on an AOD, or the derivation in question
does not include the ExtendedJetCommon module.


Former-commit-id: e19ae69281b28af218ddcbc5a9444184f638c07b