manual sweep of !11323 from 21.3 to master - Fix hard-coded values in sTGC Readout Geometry
See merge request !11323

Some values used to build the sTGC readout geometry were hard-coded
or depended on hard coded values. This commit modifies the geometry
xml and adds the extra defining parameters as well as the AGDD
tools required to access the values and remove the bad coding
practices. Two new xml files were added with the same additions.
The first, stations.v1.78 is used as the soft-link. At the time of
this commit, we are in the transition between two geometry formats,
but are still using version1. Version2 will be used in the short
future when the database will be updated.


Former-commit-id: 30b23ac8163c0278db56961934876bf2c341c188