Fix for ATLASRECTS-4459
This adds protection against the possibility that the covariance of q/p could be negative for ME tracks: in such a case, the muon is treated as if it had no ME track.
This solves the FPE reported in ATLASRECTS-4459.  Investigations into the underlying cause of this obviously incorrect result are continuing.
FT0 violation is expected to be minimal: results of RunTier0Test are available at /afs/cern.ch/work/w/wleight/public/MuonSW/ATLASRECTS-4459/, and none is observed, as expected.


Former-commit-id: b5eb112d2b1041376e0a193d2f4f52125fe07ca2