AthAllocators: Fix potential race in ArenaAllocatorRegistry::registerCreator.
ArenaAllocatorRegistry::registerCreator manages a registry of allocator
creator objects, mapping from name to an integer id.  The code was assuming
that this is only called after one has already checked that the name doesn't
exist in the registry.  However, it's possible for this to fail in MT
if two threads are registering the same creator.  Be tolerant of this.

Former-commit-id: d6d95c5e9bcd2291ba48a91fa4f2fda8acb25960