LArRawConditions: Fix race in LArConditionsContainerDB.
LArConditionsContainerDB was caching iterators in get().
This was not thread-safe, and was leading to incorrect results in MT jobs.
Fixed by removing the cache.
Tried to mitigate the performance implications by introducing a
hash map as well to use for lookups.  (We still need to keep an ordered
version for iteration to work.)


Former-commit-id: dd3c0be375e4723e57709fd2701be3a278642913