adjustment of MuonTrackingGeometry to properly build NSW
The MuonTrackingGeometry package is tanscribing the volumes built
in (Muon)GeoModel to tracking surfaces used in reconstruction. One
fragile part during the translation is the decision if a volume
is treated by the StationBuilder for active volumes or by the
InertMaterialBuilder for dead structures. Which path a volume
goes is mainly decided on its position in the volume tree and on
its name, i.e. string comparisons are involved. With the most
recent NSW geometry where the main NSW mother volume now contains
both, active and dead structures, adjustments are needed to
properly assign objects to the corresponding builder. This merge
request applies the needed modifications including support for
additionally required surfaces in the StationBuilder, and adds
some more debugging output.

The modification was checked with RunTier0Tests.py and does not
change the output of reconstruction w/o the NSW.


Former-commit-id: d95c188a7f02ed6bba37318541796c671be23ced