CMakeLists fix for DerivationFrameworkSUSY
This MR implements a fix to the CMakeLists of DFSUSY, kindly provided by Attila.
The discussion can be found here:
https://groups.cern.ch/group/hn-atlas-offlineSWHelp/Lists/Archive/DispForm.aspx?ID=28663
A suggestion from Christos was also to drop the offending include of xAODParticleEvent/CompositeParticle.h in SUSYIDWeight.cxx.
This was done, and consequently, xAODParticleEvent was dropped from CMakeLists.

I ran over a SUSY xAOD, the SUSY weights appear in the DAOD.

Thanks!
Bertrand


Former-commit-id: 4dd38b4ba5f3d234c02e1c00a1bca1ec8c16190d