Add isTestBeam property to LArBarrelPresamplerCalculator
Whether we are running a test-beam job or not should be known at
configuration time, so it is much more efficient to set a configurable
property at this point rather than checking whether we are using
a test beam geometry *every single step* in the pre-sampler. Part of
the changes for ATLASSIM-2643. (Does not change the output in my local tests.)


Former-commit-id: 94d71c38165cec8d43a49c10adf7610d5cb04c8a