Add isTestBeam property to LArBarrelPresamplerGeometry
Whether we are running a test-beam job or not should be known at
configuration time, so it is much more efficient to set a configurable
property at this point rather than checking whether we are using
a test beam geometry *every single step* in the pre-sampler. Part of
the changes for ATLASSIM-2643. (Does not change the output in my local tests.)


Former-commit-id: ffc9ec29d05010b853bca289ddb799d2e448ed46