adding valgrind suppression files
On the twiki page describing the use of valgrind with Athena

https://twiki.cern.ch/twiki/bin/viewauth/AtlasComputing/UsingValgrind

it is recommended to use several suppression files. However, some
of the suppression files are not available any longer as they were
part of the External/AtlasValgrind package which for some reason
doesn't exist anymore. Even though parts of the suppression files
seem outdated other parts may still be applicable. The suppression
files added with this MR are copies of the latest ones in rel 20.7

https://gitlab.cern.ch/atlas/athena/blob/20.7/External/AtlasValgrind/share/root.supp
https://gitlab.cern.ch/atlas/athena/blob/20.7/External/AtlasValgrind/share/Gaudi.supp
https://gitlab.cern.ch/atlas/athena/blob/20.7/External/AtlasValgrind/share/AtlasValgrind.supp
https://gitlab.cern.ch/atlas/athena/blob/20.7/External/AtlasValgrind/share/valgrind-python.supp

and were tested/used in a recent job.


Former-commit-id: 61a28293f262937f3e72b9aa1221e1da735b2c8e