fix for irreproducibility and moving hash creation to base class
When the hashes are build in the AlignmentErrorTool, there was
a stable_sort of the RIO_OnTrack pointer vectors attached to
the deviations. This stable_sort is not needed as both hashes for
rotations and translations are anyway build from the same vector.
Even worse this sorting was causing an irreproducibility as the
fitter does expect the RIO_OnTrack in a specific order and the
(stable) re-arrangement based on pointer values destroyed that.
With the removal of the sorting an irreproducibility is finally
removed and the code is (more) safe now.

To make the code more neat the hash functionality was moved from
the derived deviation classes to the deviation base class to
avoid code duplication.


Former-commit-id: b3a336cd5e5eee930cd12382ea58642096867482