SimKernel passes an McEventCollection in ISimulationSvc::simulateVector calls
In order to make the data-dependencies of the ISF code more robust it makes sense
to pass the pointer to the hard-scatter `McEventCollection` from the `SimKernel`
to the `ISimulationSvc` implementations, rather than retrieving the collection
independently later.

To this end, the prototypes of the `ISimulationSvc::simulate(Vector)` methods have
been updated to take an `McEventCollection` pointer as an argument.

The hard-scatter `McEventCollection` is then given as an argument to all
`ISimulationSvc::simulateVector` calls from `SimKernel::execute`.


Former-commit-id: 6082043e672b74eed5f94420ff1da53f20e6326b