Add functionality for r-t relation smearing (ATLASRECTS-3882)
Two services were added that can smear the t0 and r(t) values.
Using them is optional and can be steered with new job options.

The tools themselves have some configuration options, mainly the
maginitude of the smearing. The corresponding calibration files
are requested to be put into the DATASET dir (ATLINFR-2490).


Former-commit-id: 4d6ad18af4418487dde2925a59731d7abc825f38