comment out recordAddress; add stop-start of CoreDumpSvc
recordAddress for EventInfo proxy fails because it is already recorded
by the input svc. It seems everything works without recordAddress in
the event loop mgr.

CoreDumpSvc gets the number of slots from
Gaudi::Concurrency::ConcurrencyFlags::numConcurrentEvents
This is not yet set when the service is started, because we initialise
the scheduler only after fork. Therefore, we need to stop and start
CoreDumpSvc after initialising the scheduler, to pick up the number
of slots correctly.

Former-commit-id: ff04aee8c9b5a0dbd193637ed3c53deecc943b0d