skip production of AGDD outputs of Amdcsimrec from DB
For the comparison of the database content and what is produced
from a local file in order to build a geometry the Amdcsimrec
software is run. In the past the checks implemented in the package
made much sense, but today the AGDD blob is built by MuonAGDD and
the checks in Amdcsimrec are less reasonable. In the context of
ATLASSIM-3636 this MR comments out these checks. In a separate
one they will be removed (but just in master).


Former-commit-id: 97706e9f87eeb1d15310e9b1fadde3188338e7ed