cleaning up Amdcsimrec to remove obsolete code
In the far past dead structures in the muon spectrometer were built
by MuonGeoModel from dimensions stored in the Oracle tables ABRT,
AECT, ECST, FEET, GENV, JDSH, JFSH, JTSH and MBAP. The tables were
filled by Amdcsimrec. However, already before run-1 this mechanism
was disabled and the structures were built using AGDD. Therefore
Amdcsimrec can be cleaned as also MuonGeoModel got already cleaned
a while ago.

With the removal of Muonboy the AGDD mechanism became basically
decoupled from Amdcsimrec. The main migration is discussed on
ATLASSIM-3636 which links the MRs to fill the database directly
with AGDD w/o using Amdcsimrec. Only the parsing is currently
still part of Amdcsimrec, but could be decoupled any time.
In any case the table writing can be removed from Amdcsimrec.

Last but not least AgddXMLVariables can be removed as it's
not clear if that was ever used anyway.


Former-commit-id: 843c9514038da6c3375230d01858a043c7b47d36