RNN tau triggers unleashed
This MR brings the final improvement to 2018 tau triggers: RNN identification and recovery of 0p taus.
The medium RNN ID working points are set such that the rates of mediumRNN_tracktwoMVA triggers are comparable to those of medium1_tracktwoEF triggers which use BDT identification.
The RNN ID target efficiencies are 65% for 0p, 96.5% for 1p and 86.5% for 3p.
The RNN triggers allow 0-track candidates on the whole pt range, while 0p were rejected below 200 GeV for BDT triggers.
Medium RNN ID is relaxed to loose ID above 280 GeV instead of 330 GeV for BDT triggers, thanks to the increased rejection of RNN at high pt.
A minor change in the MVA energy scale code (MvaTESEvaluator.cxx) was necessary to accommodate a change in variable name in the new training, but is totally harmless.
The delta(z0) cut for tracktwoMVA was tightened from 1 to 0.75mm to recover a bit of efficiency for 3p at high mu, given that the BDT track classification cannot be used (rates too large).
Updated tunings for MVA TES and RNN ID (json training files as well as RNN score flattening ROOT files) are used.
Their transfer to cvmfs has been requested in ATLINFR-2499.

I've set the 'urgent' label because we'd like to proceed with the 21.1 validation asap, to ensure we're ready for the post-TS1 high-mu data.

Thanks!
Bertrand


Former-commit-id: e84f5e3448b602f4e12469fc14565c2d40b48926