Fixing stealth Tier-0 violation caused by DRAW_ZMUMU production
It is observed that since 21.0.72 RAWtoALL jobs produce AODs with DAOD_level muon decorations in them, leading to downstream issues. It turns out that these are being added by the DRAW_ZMUMU, which schedules the calculation of these variables and commits them to SG to enable the use of the ExpressionEvaluation to select the events. However, since the Muon output items list wildcards all MuonAux variables, these DAOD level variables, which are intended by the authors to be transient, get picked up and written into StreamAOD. This is trivially prevented by explicitly removing them from the item list in the output items script. This is put into effect in this MR.


Former-commit-id: cf128768eb003c3bd0d7155387cdf401e928ca65