no Hough seeded segment finding to make MuGirl run for NSW
In order to get MuGirl running it was necessary to turn off the
Hough seeding for the segment finding. Furthermore the clusters
in the NSW chambers need to be calibrated. Finally the cluster
segment finding was required to make aware of the NSW clusters.

In particular switching off the Hough seeding should be reviewed
and the likely resulting performance penatly is worth to be
investigated.


Former-commit-id: 55aaabeb12a250a25f2ffa5470dd99f7ce433371