Fix for non-standard StoreGateSvc names in VarHandleKey.cxx
While debugging ATLASSIM-3760 it was noticed that when `VarHandleKeys`
were used with non-standard `StoreGateSvc` names, then the resulting
`VarHandles` pointed at the defeult `StoreGateSvc_Impl` instance rather
than the instance indicated in the `VarHandleKey`.

The change in this commit resolves this issue, but in the longer term
a different solution may be adopted.


Former-commit-id: a90592d1ef08ab272f57c744d2117e1c3b3dc8df