Fix SplitPileUpToolsAlg and add SplitSFPileUpToolsAlg
Adding the missing conditions folder configuration to the
`PixelDigitizationToolHS` `CfgGetter` method. This fixes
running digitization with the `SplitPileUpToolsAlg`.
This runs full digitization for both pile-up and hard-scatter.
Then added a new configuration called `SplitSFPileUpToolsAlg`
which runs full digitization for hard-scatter and fast digitization
for pile-up.

These changes are part of the work for ATLASSIM-3737.

Former-commit-id: 86cde12241b0399c48762844a0a86dde93e6b5be