TrigPSCPythonSetup: add a possibility to run from pickle file
Added a possibility to specify a pickle file as job options and read
the full configuration from there instead of reading the python configuration.
It "works" as in it loads the configuration from the pickle file, however
it's not really tested because at this point we don't have any online job
configuration available in a pickle file.


Former-commit-id: ade99db8dd61abb89204de561975a1fef48a4bfc