Hide storegate accesses for MCTruthSimAlgs
For the MT Digi migration, we have needed to hide all the direct storegate
accesses before being able to then do the subevent accesses in a thread
safe way.  This has been done for all the other subsystem digitizations,
and this implementation takes advantage of previous extensions to the
pileup merge service.


Former-commit-id: eb110d5c9cc9159a2bb7af834990aa73ed7d097d