introduce an algorithm for overlap removal
It's more or less a basic wrapper around the IOverlapRemovalTool.
I've foregone the OR-ToolBox, which is essentially a C++ configuration
helper, which doesn't seem to be working too well with the python
configuration, i.e. I can either use the configuration via ToolBox or
configure OverlapRemovalTool via python, but not both.


Former-commit-id: 599589b4bdf64c31d27ab2a9bf17aa203529b040